use a custom TestCase.h rather than Test.h
