+++ UnitTest/pch.h
@@ -19,16 +19,14 @@
 #include <math.h>
 #include <assert.h>
 
-#include "Test.h"
-
 #include <array>
 #include <string>
-
 #include <std/string>
 #include <std/span>
 #include <std/optional>
 #include <std/cstddef>
 #include <std/bit>
+
 #include "gsl/gsl.h"
 #include "import/sys.h"
 #include "import/math.h"
@@ -36,7 +34,8 @@
 #include "import/str.h"
 #include "import/except.h"
 #include "import/mem.h"
-
 #include "str/EncodedString.h"
+
+#include "TestCase.h"
 
 #endif //PCH_H
